split just before func_us_801D542C to fix .rodata 4 byte offset
