+++ config/splat.us.stno0.yaml
@@ -79,6 +79,7 @@ segments:
       - [0x414EC, .rodata, e_diplocephalus]
       - [0x415B8, .rodata, e_skelerang]
       - [0x41624, .rodata, e_plate_lord]
+      - [0x416A0, .rodata, e_plate_lord_2]
       - [0x416F0, .rodata, e_marionette]
       - [0x4171C, .rodata, e_slinger]
       - [0x41738, .rodata, e_stone_rose]
@@ -112,6 +113,7 @@ segments:
       - [0x4F4A8, c, e_diplocephalus]
       - [0x5191C, c, e_skelerang]
       - [0x52374, c, e_plate_lord]
+      - [0x5542C, c, e_plate_lord_2]
       - [0x55E4C, c, e_ghost]
       - [0x56254, c, e_marionette]
       - [0x575E4, c, e_slinger]
